+++ done/ft_strlcpy.c
@@ -24,8 +24,8 @@ size_t	ft_strlen(const char *s)
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t len)
 {
-	int		i;
-	char	*s;
+	size_t		i;
+	char		*s;
 
 	s = (char *)src;
 	i = 0;
