+++ done/ft_strnstr.c
@@ -25,8 +25,8 @@ size_t	ft_strlen(const char *s)
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
